오타 수정

+++ src/threads/thread.h
@@ -112,11 +112,11 @@ struct thread
     /* Owned by thread.c. */
     unsigned magic;                     /* Detects stack overflow. */
 
-    int64_t wakeup_ticks;  // choi
+    int wakeup_ticks;  // choi
 
     // advanced scheduler에서 사용
-    int64_t nice;
-    int64_t recent_cpu;
+    int nice;
+    int recent_cpu;
   };
 
 /* If false (default), use round-robin scheduler.
